added new sector control scripts
Added a reward system and a "spawn" system that monitors ALiVE virtualized entities and spawns a unit to represent the fact that they are present in the sector. if different factions are in the sector, they will battle and eventually, whichever side wins, that unit will remain. if any side leaves the sector, the unit is despawned.

add 3 new folders with readmes in each explaining everything. respawn, reward_system, and spawn.